weaviate-server/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
Authlib==1.3.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
grpcio==1.64.0
grpcio-health-checking==1.64.0
grpcio-tools==1.64.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
protobuf==5.27.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.32.2
sniffio==1.3.1
typing_extensions==4.12.0
urllib3==2.2.1
validators==0.28.1
weaviate-client==4.6.3
Werkzeug==3.0.3
